aider: Added YouTube video above the first horizontal line in website.html.
# Aider chat conversation:

USER: I want to add this to website.html above the first horizontal line.  It's a youtube video I'm trying to embed- https://www.youtube.com/embed/h2-petWJw_U

ASSISTANT: To embed a YouTube video in your HTML file, you can use the `<iframe>` tag. The `src` attribute of the `<iframe>` tag will contain the URL of the YouTube video you want to embed.

Here is the code you need to add:

```html
<iframe width="560" height="315" src="https://www.youtube.com/embed/h2-petWJw_U" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
```

This code will create an iframe with a width of 560 pixels and a height of 315 pixels. The `src` attribute contains the URL of the YouTube video. The `frameborder` attribute is set to 0 to remove the border around the iframe. The `allow` attribute is used to specify what the iframe is allowed to do. The `allowfullscreen` attribute is used to allow the video to be played in fullscreen mode.

Here is the *edit block* for the change:

```html
website.html
<<<<<<< HEAD
    <div>
        <img src="horizontalline4.gif" alt="Horizontal Line">
=======
    <div>
        <iframe width="560" height="315" src="https://www.youtube.com/embed/h2-petWJw_U" frameborder="0" allow="accelerometer; autoplay; clipboard-write; encrypted-media; gyroscope; picture-in-picture" allowfullscreen></iframe>
        <img src="horizontalline4.gif" alt="Horizontal Line">
>>>>>>> updated
```

This *edit block* adds the iframe just before the first horizontal line image in the `div` tag.