aider: Added z-index and position properties to the iframe element in style.css.
# Aider chat conversation:

USER: I think we need to make sure the video is on the frontmost layer in `style.css`

ASSISTANT: To ensure that the video is on the frontmost layer, we can use the CSS `z-index` property. The `z-index` property specifies the stack order of an element. An element with greater stack order is always in front of an element with a lower stack order.

We can target the iframe containing the video in the CSS file and set its `z-index` to a high value to ensure it's on the frontmost layer.

Here is the *edit block* for the change:

```css
style.css
<<<<<<< HEAD
img {
    display: block;
    margin-right: auto;
    z-index: 9999;
    position: relative;
}
.website-image {
    z-index: 9999;
}
=======
img {
    display: block;
    margin-right: auto;
    z-index: 9999;
    position: relative;
}
.website-image {
    z-index: 9999;
}
iframe {
    z-index: 10000;
    position: relative;
}
>>>>>>> updated
```

This *edit block* adds a new rule for `iframe` elements, setting their `z-index` to 10000, which is higher than the `z-index` of the images, ensuring that the video will be on the frontmost layer. The `position: relative;` is necessary for `z-index` to take effect.